+++ leb-ref/components/interfaces/IRandomNumberRangeOptions.ts
@@ -1,6 +1,6 @@
 import INumberRangeOptions from './INumberRangeOptions';
 import IRandomRangeOptions from './IRandomRangeOptions';
 
-export default interface IRandomNumberOptions extends IRandomRangeOptions<INumberRangeOptions> {
+export default interface IRandomNumberRangeOptions extends IRandomRangeOptions<INumberRangeOptions> {
   float: boolean;
 }