+++ leb-ref/components/types/RangeOptionTypes.ts
@@ -10,7 +10,8 @@ type StringRangeOptionsT = Interfaces.IStringRangeOptions;
 type MergeRangeOptionsT = Interfaces.IMergeRangeOptions;
 type ZipRangeOptionsT = Interfaces.IZipRangeOptions;
 type RandomRangeOptionsT<T> = Interfaces.IRandomRangeOptions<T>;
-type RandomNumberOptionsT = Interfaces.IRandomNumberOptions;
+type RandomNumberRangeOptionsT = Interfaces.IRandomNumberRangeOptions;
+type RandomCharRangeOptionsT = Interfaces.IRandomCharRangeOptions;
 
 type OptionsType<T> = T extends RangeOptionsT<infer EdgeT, infer ItemT, infer DateT> ? T : never;
 type OptionsEdgeType<T> = T extends RangeOptionsT<infer EdgeT, infer ItemT, infer DateT> ? EdgeT : never;
@@ -33,5 +34,6 @@ export {
   MergeRangeOptionsT,
   ZipRangeOptionsT,
   RandomRangeOptionsT,
-  RandomNumberOptionsT,
+  RandomNumberRangeOptionsT,
+  RandomCharRangeOptionsT,
 }